Make creation of testing utils consistent
