+++ test/mocks/mockExpress.ts
@@ -1,0 +1,41 @@
+import type { Request, Response } from 'express'
+import { Order } from '../../server/models/Order'
+
+export const createMockRequest = (order?: Order): Request => {
+  return {
+    // @ts-expect-error stubbing session
+    session: {},
+    query: {},
+    params: {
+      orderId: '123456789',
+    },
+    user: {
+      username: '',
+      token: '',
+      authSource: '',
+    },
+    order,
+  }
+}
+
+export const createMockResponse = (): Response => {
+  // @ts-expect-error stubbing res.render
+  return {
+    locals: {
+      user: {
+        username: 'fakeUserName',
+        token: 'fakeUserToken',
+        authSource: 'nomis',
+        userId: 'fakeId',
+        name: 'fake user',
+        displayName: 'fuser',
+        userRoles: ['fakeRole'],
+        staffId: 123,
+      },
+    },
+    redirect: jest.fn(),
+    render: jest.fn(),
+    set: jest.fn(),
+    send: jest.fn(),
+  }
+}
